+++ src/schick/seg048.h
@@ -5,7 +5,7 @@ namespace M302de {
 /* 0x20 */
 void status_select_hero(void);
 /* 0x25 */
-void status_menu(signed short);
+void status_menu(signed int);
 /* 0x2a */
 void reset_item_selector(void);
