+++ src/schick/seg053.h
@@ -6,7 +6,7 @@ namespace M302de {
 void do_healer(void);
 
 /* 0x25 */
-signed short is_hero_healable(const struct struct_hero*);
+signed int is_hero_healable(const struct struct_hero*);
 
 #if !defined(__BORLANDC__)
 }
